chore: review comments from Anton
