+++ src/components/DebtForm/DebtForm.module.css
@@ -88,7 +88,7 @@
 
   .addDebtlabel {
     font-size: xx-large;
-    color: rgb(0, 0, 0);
+    color: 'black';
     display: flex;
     margin-top: 20px;
     margin-bottom: 20px;
