+++ src/components/Transactions/TransactionItem.js
@@ -1,33 +1,36 @@
-import { auth } from '../../services/firebase';
-import useHomeApi from '../../hooks/useHomeApi';
-import { useState, useEffect } from 'react';
-import * as React from 'react';
-import { styled } from '@mui/material/styles';
-import Card from '@mui/material/Card';
-import CardHeader from '@mui/material/CardHeader';
-import CardMedia from '@mui/material/CardMedia';
-import CardContent from '@mui/material/CardContent';
-import CardActions from '@mui/material/CardActions';
-import Collapse from '@mui/material/Collapse';
-import Avatar from '@mui/material/Avatar';
-import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import { deepPurple } from '@mui/material/colors';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { auto } from '@popperjs/core';
-import { ArrowForward } from '@mui/icons-material';
-import Button from '@mui/material/Button';
-import ButtonGroup from '@mui/material/ButtonGroup';
-const ExpandMore = styled((props) => {
+import { auth } from "../../services/firebase";
+import useHomeApi from "../../hooks/useHomeApi";
+import { useState } from "react";
+import * as React from "react";
+import { formatDate } from "../../utils/formatDate";
+import { styled } from "@mui/material/styles";
+import {
+  Card,
+  CardHeader,
+  CardContent,
+  CardActions,
+  Collapse,
+  Avatar,
+  IconButton,
+  Typography,
+  colors,
+  ButtonGroup,
+  Button,
+} from "@mui/material";
+import { ExpandMore, ArrowForward} from '@mui/icons-material'
+import { auto } from "@popperjs/core";
+
+const ExpandMoreIcon = styled((props) => {
   const { expand, ...other } = props;
   return <IconButton {...other} />;
 })(({ theme, expand }) => ({
-  transform: !expand ? 'rotate(0deg)' : 'rotate(180deg)',
-  marginLeft: 'auto',
-  transition: theme.transitions.create('transform', {
+  transform: !expand ? "rotate(0deg)" : "rotate(180deg)",
+  marginLeft: "auto",
+  transition: theme.transitions.create("transform", {
     duration: theme.transitions.duration.shortest,
   }),
 }));
+
 const TransactionItem = ({
   sender,
   receiver,
@@ -55,31 +58,20 @@ const TransactionItem = ({
     setExpanded(!expanded);
   };
 
-  const formatDate = (ISOStringDate) => {
-    const readableDate = new Date(ISOStringDate);
-    let day = readableDate.getDate();
-    let month = readableDate.getMonth() + 1;
-    let year = readableDate.getFullYear();
-    if (day < 10) {
-      day = `0` + day;
-    }
-    if (month < 10) {
-      month = `0` + month;
-    }
-    return day + `/` + month + `/` + year;
-  };
   const putTransactionsDeclineHandler = () => {
     setIsLoading(true);
     putTransactionsDecline(JSON.stringify([id]))
       .then(() => reFetchTransactions())
       .finally(() => setIsLoading(false));
   };
+
   const putTransactionsCompleteHandler = () => {
     setIsLoading(true);
     putTransactionsComplete(JSON.stringify([id]))
       .then(() => reFetchTransactions())
       .finally(() => setIsLoading(false));
   };
+
   const putTransactionsApproveHandler = () => {
     putTransactionsApprove(JSON.stringify([id]))
       .then(() => reFetchTransactions())
@@ -87,14 +79,14 @@ const TransactionItem = ({
   };
 
   return (
-    <Card sx={{ marginBottom: 2, fontSize: 'small' }}>
+    <Card sx={{ marginBottom: 2, fontSize: "small" }}>
       <CardHeader
         sx={{ margin: auto }}
         avatar={
           <>
             <Avatar
               sx={{
-                bgcolor: deepPurple[500],
+                bgcolor: colors.deepPurple[500],
                 fontSize: 12,
                 width: 66,
                 height: 66,
@@ -109,7 +101,7 @@ const TransactionItem = ({
             />
             <Avatar
               sx={{
-                bgcolor: deepPurple[500],
+                bgcolor: colors.deepPurple[500],
                 fontSize: 12,
                 width: 66,
                 height: 66,
@@ -157,14 +149,14 @@ const TransactionItem = ({
               </Button>
             )}
         </ButtonGroup>
-        <ExpandMore
+        <ExpandMoreIcon
           expand={expanded}
           onClick={handleExpandClick}
           aria-expanded={expanded}
           aria-label="show more"
         >
-          <ExpandMoreIcon />
-        </ExpandMore>
+          <ExpandMore />
+        </ExpandMoreIcon>
       </CardActions>
       <Collapse in={expanded} timeout="auto" unmountOnExit>
         <CardContent>
@@ -180,5 +172,5 @@ const TransactionItem = ({
     </Card>
   );
 };
-// A|-|T0N пеDICK
+
 export default TransactionItem;
