+++ src/components/Transactions/TransactionsList.js
@@ -1,9 +1,10 @@
-import useHomeApi from '../../hooks/useHomeApi';
-import { useEffect, useMemo, useState } from 'react';
-import TransactionItem from './TransactionItem';
-import classes from './TransactionsList.module.css';
-import Spinner from '../Spinner/Spinner';
-import { Card, List } from '@mui/material';
+import useHomeApi from "../../hooks/useHomeApi";
+import { useEffect, useMemo, useState } from "react";
+import TransactionItem from "./TransactionItem";
+import classes from "./TransactionsList.module.css";
+import Spinner from "../Spinner/Spinner";
+import { Grid, Box } from "@mui/material";
+
 const TransactionsList = () => {
   const [transactions, setTransactions] = useState([]);
   const [users, setUsers] = useState([]);
@@ -28,23 +29,24 @@ const TransactionsList = () => {
       (obj1, obj2) => new Date(obj2.date) - new Date(obj1.date)
     );
     return sortedTransactions.map((transaction) => (
-      <TransactionItem
-        key={transaction.id}
-        id={transaction.id}
-        sender={formatUserName(transaction.sender)}
-        receiver={formatUserName(transaction.receiver)}
-        currency={transaction.currency}
-        amount={transaction.amount}
-        description={transaction.description}
-        date={transaction.date}
-        status={transaction.status}
-        created={transaction.created}
-        updated={transaction.updated}
-        senderId={transaction.sender}
-        recieverId={transaction.receiver}
-        reFetchTransactions={reFetchTransactions}
-        
-      />
+      <Grid item xs={12} md={6} lg={4}>
+        <TransactionItem
+          key={transaction.id}
+          id={transaction.id}
+          sender={formatUserName(transaction.sender)}
+          receiver={formatUserName(transaction.receiver)}
+          currency={transaction.currency}
+          amount={transaction.amount}
+          description={transaction.description}
+          date={transaction.date}
+          status={transaction.status}
+          created={transaction.created}
+          updated={transaction.updated}
+          senderId={transaction.sender}
+          recieverId={transaction.receiver}
+          reFetchTransactions={reFetchTransactions}
+        />
+      </Grid>
     ));
   }, [transactions]);
 
@@ -55,9 +57,11 @@ const TransactionsList = () => {
   }
 
   return (
-    <Card>
-      <List>{transformedTransactions}</List>
-    </Card>
+    <Box sx={{ flexGrow: 1 }}>
+      <Grid container spacing={2}>
+        {transformedTransactions}
+      </Grid>
+    </Box>
   );
 };
