+++ src/components/Transactions/TransactionsList.module.css
@@ -1,48 +1,4 @@
-.transactions {
-    max-width: 60rem;
-    width: 90%;
-    margin: 2rem auto;
-    animation: transactions-appear 1s ease-out forwards;
-  }
-  
-  .transactionsLoading {
-    text-align: center;
-    color: white
-  }
-  
-  .transactionsError {
-    text-align: center;
-    color: red;
-  }
-  
-  .transactions ul {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-  }
-  
-  @keyframes transactions-appear {
-    from {
-      opacity: 0;
-      transform: translateY(3rem);
-    }
-  
-    to {
-      opacity: 1;
-      transform: translateY(0);
-    }
-  }
-  .card {
-    padding: 1rem;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.25);
-    border-radius: 14px;
-    background-color: white;
-  }
-
-  .noTransactionsAvailable {
-    font-size: xx-large;
-    color: red;
-  }
-
-
-  
+.noTransactionsAvailable {
+  font-size: xx-large;
+  color: red;
+}
