+++ src/utils/formatDate.js
@@ -1,0 +1,13 @@
+export const formatDate = (ISOStringDate) => {
+  const readableDate = new Date(ISOStringDate);
+  let day = readableDate.getDate();
+  let month = readableDate.getMonth() + 1;
+  let year = readableDate.getFullYear();
+  if (day < 10) {
+    day = `0` + day;
+  }
+  if (month < 10) {
+    month = `0` + month;
+  }
+  return day + `/` + month + `/` + year;
+};
